+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii 2 Content Management extension Change Log
 1.0.1 under development
 -----------------------
 
+- Bug #1: Fixed `Manager::getMetaData()` produces E_WARNING for the item, which is not present at source storage (klimov-paul)
 - Enh: Usage of deprecated `yii\base\InvalidParamException` replaced by `yii\base\InvalidArgumentException` (klimov-paul)