+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Yii 2 Content Management extension Change Log
 =============================================
 
-1.0.1 under development
------------------------
+1.0.1, July 02, 2018
+--------------------
 
 - Bug #1: Fixed `Manager::getMetaData()` produces E_WARNING for the item, which is not present at source storage (klimov-paul)
 - Enh #2: Created `yii2tech\content\mail\MailerContentBehavior` for mail composition simplification (klimov-paul)